# programming-lng-service/src/swagger/programming-api.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "programming-language-service",
        "description": "This is a sample programming language microservice",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "localhost"
        }
    ],
    "paths": {
        "/api/programming/{id}": {
            "get": {
                "tags": ["programming"],
                "summary": "get a programming language by id",
                "operationId": "getProgrammingLanguage",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProgrammingLanguageData"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "unsuccesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Errors"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/programming/all": {
            "get": {
                "tags": ["programming"],
                "summary": "retrieve all programming language from database",
                "operationId": "getAllProgrammingLanguage",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProgrammingLanguageData"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "unsuccesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Errors"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/programming/learning": {
            "post": {
                "tags": ["programming"],
                "summary": "retrieve learning statsu of a skill database",
                "operationId": "getStatus",
                "requestBody": {
                    "description": "add programming language to the backend",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "number"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "unsuccesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Errors"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/programming/destroy": {
            "post": {
                "tags": ["programming"],
                "summary": "delete a programming language by id",
                "operationId": "deleteTask",
                "requestBody": {
                    "description": "delete programming language from the backend",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "unsuccesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Errors"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/programming/add": {
            "post": {
                "tags": ["programming"],
                "summary": "add a programming language",
                "operationId": "addProgrammingLanguage",
                "requestBody": {
                    "description": "add programming language to the backend",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProgrammingLanguageData"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "unsuccesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Errors"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/programming/update": {
            "post": {
                "tags": ["programming"],
                "summary": "edit a programming language",
                "operationId": "EditProgrammingLanguage",
                "requestBody": {
                    "description": "edit programming language data in the backend",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProgrammingLanguageData"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "unsuccesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Errors"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "programming language id values that need to be considered for filter",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        },
        "schemas": {
            "Errors": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "field": {
                        "nullable": true,
                        "type": "string"
                    }
                }
            },
            "ProgrammingLanguageData": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "version": {
                        "type": "number"
                    },
                    "dbStatus": {
                        "type": "string"
                    },
                    "course": {
                        "$ref": "#/components/schemas/Course"
                    },
                    "books": {
                        "$ref": "#/components/schemas/Book"
                    }
                }
            },
            "Course": {
                "type": "string",
                "nullable": true
            },
            "Book": {
                "type": "string",
                "nullable": true
            }
        }
    }
}